sme/requirements.txt
Cheetah==2.4.4
Flask==0.10.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.10
SQLAlchemy==1.0.12
SecretStorage==2.0.0
WTForms==2.1
Werkzeug==0.11.4
apt-xapian-index==0.45
argparse==1.2.1
bzr==2.7.0dev1
chardet==2.0.1
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
gevent==1.0.2
greenlet==0.4.9
html5lib==0.999
httplib2==0.8
iotop==0.6
itsdangerous==0.24
jsonpatch==1.3
jsonpointer==1.0
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lxml==3.3.3
meld3==0.6.10
mercurial==2.8.2
oauth==1.0.1
paramiko==1.10.1
prettytable==0.7.2
pycrypto==2.6.1
pygobject==3.12.0
pygpgme==0.3
pyserial==2.6
python-apt==0.9.3.5ubuntu2
python-debian==0.1.21-nmu2ubuntu2
python-docx==0.8.5
requests==2.2.1
retinasdk==1.0.0
simplejson==3.3.1
six==1.5.2
ssh-import-id==3.21
supervisor==3.0b2
urllib3==1.7.1
wadllib==1.3.2
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5